[LCR-14] refactor: CustomUserDetailsService로 클래스명 및 패키지 경로 변경
